fix: 🐛 Fetch legs from chain while instruction affirm/execution
`Instruction.getLegs` was modified to use middleware to fetch the leg
information. This caused the affirm/execution procedure to wait for the
instruction to be synced via middleware before affirming/executing (when
middleware was available).
